+++ MatrixSDK/MXRestClient.m
@@ -2,6 +2,7 @@
  Copyright 2014 OpenMarket Ltd
  Copyright 2017 Vector Creations Ltd
  Copyright 2018 New Vector Ltd
+ Copyright 2019 The Matrix.org Foundation C.I.C
 
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -3713,8 +3714,14 @@ MXAuthAction;
 
 #pragma mark - Certificates
 
--(void)setPinnedCertificates:(NSSet <NSData *> *)pinnedCertificates {
+-(void)setPinnedCertificates:(NSSet <NSData *> *)pinnedCertificates
+{
     httpClient.pinnedCertificates = pinnedCertificates;
+}
+
+- (void)setPinnedCertificates:(NSSet<NSData *> *)pinnedCertificates withPinningMode:(MXHTTPClientSSLPinningMode)pinningMode
+{
+    [httpClient setPinnedCertificates:pinnedCertificates withPinningMode:pinningMode];
 }
 
 #pragma mark - VoIP API